Deal with 4 小韻s without valid 反切
Entries in these 4 小韻s are technically erroneous versions of other
characters (the very reason why their shape and 反切 do not make sense),
but are nonetheless provided a way to input.

- 𤜼 is given the same 地位 as its correct form "犳"
- 𡰝 and 㶒 are given more reasonable 地位s from their 反切 (taken
  account of mistakenly written characters in the 反切)
- 𣅝's 地位 is still from its 反切

In addition, "突" is added as a variant of "宊" (but not the other way
around, as "宊"(透母) can be recognized as a form of "突" in some
sources (e.g. 集韻) but "突"(定母) is typically not written as "宊").